+++ frontend/src/utils/testing/fixtures.ts
@@ -29,6 +29,8 @@ import {
   UserRole,
 } from "src/types/userTypes";
 
+import { FormattedFormValidationWarning } from "src/components/applyForm/types";
+
 export const mockOpportunity: BaseOpportunity = {
   opportunity_id: "63588df8-f2d1-44ed-a201-5804abba696a",
   legacy_opportunity_id: 12345,
@@ -727,4 +729,14 @@ export const mockApplicationSubmission: ApplicationSubmission = {
   download_path: "http://s3signedurl.com/download.zip",
   file_size_bytes: 10,
   legacy_tracking_number: 1,
+};
+
+export const fakeValidationError: FormattedFormValidationWarning = {
+  definition: "/properties/field_one",
+  field: "$.field_one",
+  formatted: "Submission Type is required",
+  htmlField: "field_one",
+  message: "'field_one' is a required property",
+  type: "required",
+  value: null,
 };
